requirements.txt
# KidGuard Dependencies

# Claude API
anthropic>=0.40.0

# Computer Vision
opencv-python>=4.8.0
face_recognition>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0

# Screen capture
mss>=9.0.0
pyautogui>=0.9.54

# Audio processing (for future)
# sounddevice>=0.4.6
# scipy>=1.11.0

# Browser automation
playwright>=1.40.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Web UI
flask>=3.0.0

# Notifications
python-telegram-bot>=20.0

# Process monitoring
psutil>=5.9.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Live Monitor
pyautogui>=0.9.54
pyperclip>=1.8.2
